+++ src/pyuvdata/uvdata/fhd.py
@@ -585,7 +585,7 @@ class FHD(UVData):
         #   because they depend on the phasing of the visibilities)
         # the values in bl_info.JDATE are the JD for each integration.
         # We need to expand up to Nblts.
-        int_times = list(utils.helpers._get_iterable(bl_info["JDATE"][0]))
+        int_times = list(utils.tools._get_iterable(bl_info["JDATE"][0]))
         bin_offset = bl_info["BIN_OFFSET"][0]
         if self.Ntimes != len(int_times):
             warnings.warn(
@@ -772,7 +772,7 @@ class FHD(UVData):
         else:
             self.history = ""
 
-        if not utils.helpers._check_history_version(
+        if not utils.history._check_history_version(
             self.history, self.pyuvdata_version_str
         ):
             self.history += self.pyuvdata_version_str
