+++ src/pyuvdata/uvdata/uvh5.py
@@ -18,7 +18,6 @@ from docstring_parser import DocstringStyle
 
 from .. import Telescope, utils
 from ..docstrings import copy_replace_short_description
-from ..utils import helpers
 from ..utils.file_io import hdf5 as hdf5_utils
 from . import UVData
 
@@ -205,7 +204,7 @@ class FastUVH5Meta(hdf5_utils.HDF5Meta):
 
     def get_blt_order(self) -> tuple[str]:
         """Get the blt order from analysing metadata."""
-        return helpers.determine_blt_order(
+        return utils.bltaxis.determine_blt_order(
             time_array=self.time_array,
             ant_1_array=self.ant_1_array,
             ant_2_array=self.ant_2_array,
@@ -252,7 +251,7 @@ class FastUVH5Meta(hdf5_utils.HDF5Meta):
         ):
             return True
 
-        is_rect, self.__time_first = helpers.determine_rectangularity(
+        is_rect, self.__time_first = utils.bltaxis.determine_rectangularity(
             time_array=self.time_array,
             baseline_array=self.baseline_array,
             nbls=self.Nbls,
@@ -514,7 +513,7 @@ class UVH5(UVData):
             proc = None
 
             if run_check_acceptability:
-                helpers.check_lsts_against_times(
+                utils.times.check_lsts_against_times(
                     jd_array=self.time_array,
                     lst_array=self.lst_array,
                     telescope_loc=self.telescope.location,
@@ -588,7 +587,9 @@ class UVH5(UVData):
         if "time_axis_faster_than_bls" in obj.header:
             self.time_axis_faster_than_bls = obj.time_axis_faster_than_bls
 
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
 
         # Optional parameters
@@ -819,7 +820,7 @@ class UVH5(UVData):
             # max_nslice_frac of 0.1 yields slice speedup over fancy index for HERA data
             # See pyuvdata PR #805
             if blt_inds is not None:
-                blt_slices, blt_sliceable = helpers._convert_to_slices(
+                blt_slices, blt_sliceable = utils.tools._convert_to_slices(
                     blt_inds, max_nslice_frac=0.1
                 )
             else:
@@ -827,7 +828,7 @@ class UVH5(UVData):
                 blt_sliceable = True
 
             if freq_inds is not None:
-                freq_slices, freq_sliceable = helpers._convert_to_slices(
+                freq_slices, freq_sliceable = utils.tools._convert_to_slices(
                     freq_inds, max_nslice_frac=0.1
                 )
             else:
@@ -835,7 +836,7 @@ class UVH5(UVData):
                 freq_sliceable = True
 
             if pol_inds is not None:
-                pol_slices, pol_sliceable = helpers._convert_to_slices(
+                pol_slices, pol_sliceable = utils.tools._convert_to_slices(
                     pol_inds, max_nslice_frac=0.5
                 )
             else:
