+++ src/pyuvdata/uvflag/uvflag.py
@@ -16,7 +16,6 @@ import numpy as np
 from .. import Telescope, UVCal, UVData
 from .. import parameter as uvp
 from .. import utils
-from ..utils import helpers
 from ..uvbase import UVBase
 
 __all__ = ["UVFlag", "flags2waterfall", "and_rows_cols"]
@@ -826,13 +825,13 @@ class UVFlag(UVBase):
 
         if run_check_acceptability:
             # Check antenna positions
-            helpers.check_surface_based_positions(
+            utils.coordinates.check_surface_based_positions(
                 antenna_positions=self.telescope.antenna_positions,
                 telescope_loc=self.telescope.location,
                 raise_error=False,
             )
 
-            helpers.check_lsts_against_times(
+            utils.times.check_lsts_against_times(
                 jd_array=self.time_array,
                 lst_array=self.lst_array,
                 telescope_loc=self.telescope.location,
@@ -1201,7 +1200,9 @@ class UVFlag(UVBase):
         self.clear_unused_attributes()
         self.history += "Pol axis collapse. "
 
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
 
         if run_check:
@@ -1323,7 +1324,9 @@ class UVFlag(UVBase):
         self._set_type_waterfall()
         self.history += 'Collapsed to type "waterfall". '  # + self.pyuvdata_version_str
 
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
 
         self.clear_unused_attributes()
@@ -1596,7 +1599,9 @@ class UVFlag(UVBase):
 
         self.history += 'Broadcast to type "baseline". '
 
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
 
         if run_check:
@@ -1761,7 +1766,9 @@ class UVFlag(UVBase):
         self._set_type_antenna()
         self.history += 'Broadcast to type "antenna". '
 
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
 
         if run_check:
@@ -1807,7 +1814,9 @@ class UVFlag(UVBase):
                 "Unknown UVFlag mode: " + self.mode + ". Cannot convert to flag."
             )
         self.history += 'Converted to mode "flag". '
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
         self.clear_unused_attributes()
 
@@ -1880,7 +1889,9 @@ class UVFlag(UVBase):
             )
         self.history += 'Converted to mode "metric". '
 
-        if not helpers._check_history_version(self.history, self.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            self.history, self.pyuvdata_version_str
+        ):
             self.history += self.pyuvdata_version_str
         self.clear_unused_attributes()
 
@@ -1946,7 +1957,7 @@ class UVFlag(UVBase):
             )
 
         # Update filename parameter
-        this.filename = helpers._combine_filenames(this.filename, other.filename)
+        this.filename = utils.tools._combine_filenames(this.filename, other.filename)
         if this.filename is not None:
             this._filename.form = (len(this.filename),)
 
@@ -2148,7 +2159,9 @@ class UVFlag(UVBase):
                 )
 
         this.history += "Data combined along " + axis + " axis. "
-        if not helpers._check_history_version(this.history, this.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            this.history, this.pyuvdata_version_str
+        ):
             this.history += this.pyuvdata_version_str
 
         this.Ntimes = np.unique(this.time_array).size
@@ -2244,7 +2257,9 @@ class UVFlag(UVBase):
         if other.history not in this.history:
             this.history += "Flags OR'd with: " + other.history
 
-        if not helpers._check_history_version(this.history, this.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            this.history, this.pyuvdata_version_str
+        ):
             this.history += this.pyuvdata_version_str
 
         if run_check:
@@ -2310,7 +2325,7 @@ class UVFlag(UVBase):
 
         """
         # Ensure others is iterable (in case of single UVFlag object)
-        # cannot use helpers._get_iterable because the object itself is iterable
+        # cannot use utils.tools._get_iterable because the object itself is iterable
         if not isinstance(others, (list, tuple, np.ndarray)):
             others = [others]
 
@@ -2342,7 +2357,9 @@ class UVFlag(UVBase):
         this.weights_array = warray
         this.history += "Combined metric arrays. "
 
-        if not helpers._check_history_version(this.history, this.pyuvdata_version_str):
+        if not utils.history._check_history_version(
+            this.history, this.pyuvdata_version_str
+        ):
             this.history += this.pyuvdata_version_str
 
         if run_check:
@@ -2460,7 +2477,7 @@ class UVFlag(UVBase):
 
         # test for blt_inds presence before adding inds from antennas & times
         if blt_inds is not None:
-            blt_inds = helpers._get_iterable(blt_inds)
+            blt_inds = utils.tools._get_iterable(blt_inds)
             if np.array(blt_inds).ndim > 1:
                 blt_inds = np.array(blt_inds).flatten()
             if self.type == "baseline":
@@ -2470,7 +2487,7 @@ class UVFlag(UVBase):
             n_selects += 1
 
         if antenna_nums is not None:
-            antenna_nums = helpers._get_iterable(antenna_nums)
+            antenna_nums = utils.tools._get_iterable(antenna_nums)
             if np.array(antenna_nums).ndim > 1:
                 antenna_nums = np.array(antenna_nums).flatten()
             if n_selects > 0:
@@ -2601,7 +2618,7 @@ class UVFlag(UVBase):
                 blt_inds = ant_blt_inds
 
         if times is not None:
-            times = helpers._get_iterable(times)
+            times = utils.tools._get_iterable(times)
             if np.array(times).ndim > 1:
                 times = np.array(times).flatten()
 
@@ -2650,14 +2667,14 @@ class UVFlag(UVBase):
             blt_inds = sorted(set(blt_inds))
 
         if freq_chans is not None:
-            freq_chans = helpers._get_iterable(freq_chans)
+            freq_chans = utils.tools._get_iterable(freq_chans)
             if np.array(freq_chans).ndim > 1:
                 freq_chans = np.array(freq_chans).flatten()
             if frequencies is None:
                 frequencies = np.squeeze(self.freq_array)[freq_chans]
 
             else:
-                frequencies = helpers._get_iterable(frequencies)
+                frequencies = utils.tools._get_iterable(frequencies)
                 frequencies = np.sort(
                     list(
                         set(frequencies) | set(np.squeeze(self.freq_array)[freq_chans])
@@ -2665,7 +2682,7 @@ class UVFlag(UVBase):
                 )
 
         if frequencies is not None:
-            frequencies = helpers._get_iterable(frequencies)
+            frequencies = utils.tools._get_iterable(frequencies)
             if np.array(frequencies).ndim > 1:
                 frequencies = np.array(frequencies).flatten()
             if n_selects > 0:
@@ -2689,7 +2706,7 @@ class UVFlag(UVBase):
             freq_inds = None
 
         if polarizations is not None:
-            polarizations = helpers._get_iterable(polarizations)
+            polarizations = utils.tools._get_iterable(polarizations)
             if np.array(polarizations).ndim > 1:
                 polarizations = np.array(polarizations).flatten()
             if n_selects > 0:
@@ -3130,7 +3147,7 @@ class UVFlag(UVBase):
                         "freq_array spacing."
                     )
                     freq_delta = np.diff(np.squeeze(self.freq_array))
-                    if helpers._test_array_constant_spacing(
+                    if utils.tools._test_array_constant_spacing(
                         self.freq_array, tols=self._freq_array.tols
                     ):
                         self.channel_width = np.full(self.Nfreqs, freq_delta[0])
@@ -3226,7 +3243,7 @@ class UVFlag(UVBase):
 
                 self.history += history
 
-                if not helpers._check_history_version(
+                if not utils.history._check_history_version(
                     self.history, self.pyuvdata_version_str
                 ):
                     self.history += self.pyuvdata_version_str
@@ -3450,7 +3467,7 @@ class UVFlag(UVBase):
                 polarization_array = self.polarization_array
             header["polarization_array"] = polarization_array
 
-            if not helpers._check_history_version(
+            if not utils.history._check_history_version(
                 self.history, self.pyuvdata_version_str
             ):
                 self.history += self.pyuvdata_version_str
@@ -3587,7 +3604,7 @@ class UVFlag(UVBase):
         if waterfall:
             self._set_type_waterfall()
             self.history += 'Flag object with type "waterfall" created. '
-            if not helpers._check_history_version(
+            if not utils.history._check_history_version(
                 self.history, self.pyuvdata_version_str
             ):
                 self.history += self.pyuvdata_version_str
@@ -3611,7 +3628,7 @@ class UVFlag(UVBase):
         else:
             self._set_type_baseline()
             self.history += 'Flag object with type "baseline" created. '
-            if not helpers._check_history_version(
+            if not utils.history._check_history_version(
                 self.history, self.pyuvdata_version_str
             ):
                 self.history += self.pyuvdata_version_str
@@ -3750,7 +3767,7 @@ class UVFlag(UVBase):
         if waterfall:
             self._set_type_waterfall()
             self.history += 'Flag object with type "waterfall" created. '
-            if not helpers._check_history_version(
+            if not utils.history._check_history_version(
                 self.history, self.pyuvdata_version_str
             ):
                 self.history += self.pyuvdata_version_str
@@ -3773,7 +3790,7 @@ class UVFlag(UVBase):
         else:
             self._set_type_antenna()
             self.history += 'Flag object with type "antenna" created. '
-            if not helpers._check_history_version(
+            if not utils.history._check_history_version(
                 self.history, self.pyuvdata_version_str
             ):
                 self.history += self.pyuvdata_version_str
